@@ -1,18 +1,18 @@
 [project]
 name = "graiax-fastapi"
-version = "0.3.0"
+version = "0.4.0"
 description = "FastAPI launcher for GraiaCommunity."
 authors = [
     { name = "BlueGlassBlock", email = "blueglassblock@outlook.com" },
     { name = "Red_lnn", email = "w731347477@gmail.com" },
 ]
 dependencies = [
-    "graia-amnesia>=0.7.1,<0.8.0",
     "fastapi>=0.101.0",
+    "graia-amnesia>=0.8.2",
     "graia-saya>=0.0.18",
-    "typing-extensions>=4.7.1,<5.0",
+    "launart>=0.7.1",
 ]
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 readme = "README.md"
 license = { text = "MIT" }
 
@@ -37,23 +37,21 @@ test = { composite = [
 [tool.pdm.dev-dependencies]
 dev = [
     "black>=23.7.0",
-    "isort>=5.12.0",
-    "launart>=0.6.4,<0.7.0",
-    "uvicorn[standard]>=0.23.2",
     "coverage>=7.2.7",
+    "graia-broadcast>=0.23.2",
+    "httpx>=0.24.1",
+    "isort>=5.12.0",
     "pytest>=7.4.0",
     "pytest-asyncio>=0.21.1",
-    "httpx>=0.24.1",
-    "graia-broadcast>=0.23.1",
 ]
 
 [tool.black]
 line-length = 120
-target-version = ["py38", "py39", "py310", "py311"]
+target-version = ["py39", "py310", "py311"]
 
 [tool.ruff]
 line-length = 120
-target-version = "py38"
+target-version = "py39"
 select = [
     # https://beta.ruff.rs/docs/rules
     "E",  # pycodestyle
